+++ src/app/roadmap/post/[[...id]]/@roadmapInfo/page.tsx
@@ -11,6 +11,7 @@ import About from './About';
 import ReactFlow from './reactFlow/ReactFlow';
 
 import { CustomEdge, CustomNode, Viewport } from '@/types/reactFlow';
+import { useRouter } from 'next/navigation';
 
 export interface RoadMapInfo extends Post {
   [key: string]: unknown;
@@ -46,6 +47,7 @@ const loadDataFromApi = async (pageParam: string) => {
 
 const Roadmap = ({ params }: { params: { id: string } }) => {
   const { id } = params;
+  const router = useRouter();
   const { data, isLoading, isError, isSuccess } = useQuery({
     queryKey: [`post${id}`],
     queryFn: async () => await loadDataFromApi(id),
@@ -54,13 +56,8 @@ const Roadmap = ({ params }: { params: { id: string } }) => {
   if (isLoading) return <div>is loading</div>;
   if (isError) return <div>oops something went wrong!🥲</div>;
   if (data === null || !data?.roadMapInfo) {
-    return (
-      <main>
-        <section>
-          <div>empty</div>
-        </section>
-      </main>
-    );
+    router.replace('/error');
+    return <></>;
   }
 
   if (isSuccess && 'roadMapInfo' in data) {
